ai-service/requirements.txt
# requirements.txt — Dépendances du microservice IA (FastAPI)
# Plateforme GED — DSI Ministère de l'Éducation Nationale

# ── Framework & serveur ──────────────────────
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
pydantic-settings==2.1.0

# ── NLP & IA ─────────────────────────────────
transformers==4.38.1
torch==2.2.0
spacy==3.7.4
sentence-transformers==2.3.1
scikit-learn==1.4.2
numpy==1.26.4

# ── OCR ──────────────────────────────────────
pytesseract==0.3.10
Pillow==10.2.0
pdf2image==1.17.0

# ── Traitement PDF ───────────────────────────
PyPDF2==3.0.1
pdfplumber==0.10.4

# ── Détection de langue ──────────────────────
langdetect==1.0.9

# ── Stockage S3 / MinIO ──────────────────────
boto3==1.34.34

# ── Utilitaires ──────────────────────────────
httpx==0.27.0
python-dotenv==1.0.1
loguru==0.7.2
